Update breakpad (432->452)
TEST=none
BUG=none
Review URL: http://codereview.chromium.org/503052

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@34916 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 7c4b495825710ab2f37967a3963ac48a977bcfc6

+++ chrome/tools/crash_service/crash_service.cc
@@ -37,7 +37,7 @@ bool CustomInfoToMap(const google_breakpad::ClientInfo* client_info,
                      const std::wstring& reporter_tag, CrashMap* map) {
   google_breakpad::CustomClientInfo info = client_info->GetCustomInfo();
 
-  for (int i = 0; i < info.count; ++i) {
+  for (uintptr_t i = 0; i < info.count; ++i) {
     (*map)[info.entries[i].name] = info.entries[i].value;
   }
